sp: Split source, add CDN auth token support
